webapp/requirements.txt
django
django_extensions
# jinja
markdown
#psycopg2
psycopg2-binary
django-environ